move get_contract_expiry_status to contract module
Move the function to a module better allow other module to reuse it
